Move all random aspects of generation in QuerySequenceTemplate to own file
