+++ lib/utils/RandomUtils.ts
@@ -1,4 +1,7 @@
+import type * as RDF from '@rdfjs/types';
+import { DataFactory } from 'rdf-data-factory';
 import type * as seedrandom from 'seedrandom';
+import { IEntityLogits } from '../QuerySequenceTemplateProvider';
 
 export interface IProbabilities<T> {
   probability: number;
@@ -47,4 +50,47 @@ export function sampleHit(rng: seedrandom.PRNG, probabilityHit: number): boolean
 
 export function randomIntFromInterval(rng: seedrandom.PRNG, min: number, max: number): number {
   return Math.floor(rng() * (max - min + 1) + min);
+}
+
+
+export function sampleVariableTerm(
+  variable: string,
+  user: string, 
+  nSamples: number,
+  variableProbabilities: Record<string, Record<string, IEntityLogits[]>>,
+  DF: DataFactory,
+  rng: seedrandom.PRNG
+): RDF.Term[] {
+  if (Object.keys(variableProbabilities[variable]).length <= nSamples) {
+    throw new Error('Trying to sample more values than there are elements');
+  }
+  const probabilities = variableProbabilities[variable];
+  if (!probabilities) {
+    throw new Error(`No probabilities found for variable '${variable}'`);
+  }
+  const logits = probabilities[user];
+  if (!logits) {
+    throw new Error(`No logits found for user '${user}' for variable '${variable}'`);
+  }
+  const sampled: string[] = [];
+  while (sampled.length < nSamples) {
+    const newSample = sampleTerm(logits, rng);
+    if (!sampled.includes(newSample)) {
+      sampled.push(newSample);
+    }
+  }
+  return sampled.map(sample => DF.namedNode(sample));
+}
+
+export function sampleTerm(logits: IEntityLogits[], rng: seedrandom.PRNG): string {
+  const r = rng();
+  let cumulative = 0;
+
+  for (const item of logits) {
+    cumulative += item.similarity;
+    if (r < cumulative) {
+      return item.entity;
+    }
+  }
+  throw new Error('Failed sampling, likely due to probabilities not summing to 1.');
 }
